requirements.txt
googletrans==3.1.0a0
diffusers 
transformers 
openai-clip
scikit-learn
ipywidgets
accelerate
torch 
streamlit
flask
torchvision